flask_server/requirements.txt
flask>=2.3.0,<4
flask-cors>=4.0.0
supabase>=2.0.0,<3
python-dotenv>=1.0.0,<2
websockets>=14.0

